Added Hal Leonard Real Book vol 3, Bb version
This file was originally a copy of the existing realbk3h.csv.  I added a column for "key", "composers", and a column that is the song title with the key in parents.  Ie. "All Blues" exists as a value, and "All Blues (Bb)" as a value.  I have charts in C and in Bb and needed way to distinguish.  _There may be better ways to do this_.  :-)